Added load more styling and hook to get params
