+++ src/app/components/LaunchCard/index.css
@@ -4,6 +4,7 @@
   padding: 1rem;
   display: flex;
   flex-direction: column;
+  gap: 1rem;
 }
 
 .launch-card-details-container {
