+++ src/app/components/LaunchList/index.css
@@ -3,4 +3,10 @@
   flex-direction: column;
   gap: 0.25rem;
   justify-content: space-evenly;
+  min-height: 40vh;
+}
+
+.launch-list-load-more {
+  width: 100%;
+  text-align: center;
 }
