+++ src/app/components/Loading/index.css
@@ -1,0 +1,8 @@
+.loading-container {
+  width: 100%;
+  padding: 4rem;
+  text-align: center;
+  font-size: 1.2rem;
+  font-style: italic;
+  color: var(--primary);
+}
