+++ src/app/components/Loading/index.tsx
@@ -1,0 +1,7 @@
+import './index.css';
+
+const Loading = () => {
+  return <div className="loading-container">Loading...</div>;
+};
+
+export default Loading;
